requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
comtypes==1.1.8
Django==3.1.4
idna==2.10
logModel==0.0.1
lxml==4.6.2
mysqldbModel==1.1.0
Pillow==8.1.2
pygame==2.0.1
PyMySQL==1.0.2
pypiwin32==223
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyttsx3==2.90
pytz==2020.5
pywin32==300
requests==2.25.1
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
